requirements.txt
# Core dependencies for Payload Generator Pecha
requests>=2.32.0
Levenshtein>=0.27.0
pydantic>=2.11.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=8.4.0

# Optional: for development
# pytest-cov>=4.0.0  # For test coverage reports